Move HIP SharedAllocationRecord to its own files
